app/requirements.txt
requests
ics
httpx
py-machineid
selectolax
pycryptodomex